requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==21.2.0
pydantic==2.8.2
pydantic-settings==2.3.4
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2
redis==5.0.7
celery==5.4.0
httpx==0.27.2
python-dotenv==1.0.1
prometheus-client==0.20.0
tenacity==9.0.0
strawberry-graphql==0.230.0
openai>=1.3.0
aiodataloader>=0.2.0